Remove the context manager for filters
I think it was a threading race condition messing things up, and simplest way around that right now is just assume that assignment is atomic in Python (which it is due to the GIL :tada:). Other than that, it is waiting for Pydantic 2 to add context to the parse calls so we can pass in the filters as context for the validators instead of having this global object around. Regrets.